server/requirements.txt
fastapi==0.120.1
openai==2.6.1
python-dotenv==1.2.1
python_magic==0.4.27
slowapi==0.1.9
uvicorn==0.38.0
